LiveQuery.kt: return the SqliteSequenceNumber of cached data
