+++ firebase-dataconnect/src/main/kotlin/com/google/firebase/dataconnect/querymgr/QueryManager.kt
@@ -18,7 +18,10 @@ package com.google.firebase.dataconnect.querymgr
 
 import com.google.firebase.dataconnect.QueryRef.FetchPolicy
 import com.google.firebase.dataconnect.core.QueryRefImpl
+import com.google.firebase.dataconnect.core.toDataSourceEnum
 import com.google.firebase.dataconnect.util.SequencedReference
+import com.google.firebase.dataconnect.util.SequencedReference.Companion.map
+import com.google.firebase.dataconnect.util.TaggedReference
 
 internal class QueryManager(private val liveQueries: LiveQueries) {
   suspend fun <Data, Variables> execute(
@@ -26,12 +29,14 @@ internal class QueryManager(private val liveQueries: LiveQueries) {
     fetchPolicy: FetchPolicy,
   ): SequencedReference<Result<DataSourcePair<Data>>> =
     liveQueries.withLiveQuery(query) {
-      it.execute(
-        dataDeserializer = query.dataDeserializer,
-        dataSerializersModule = query.dataSerializersModule,
-        callerSdkType = query.callerSdkType,
-        fetchPolicy = fetchPolicy,
-      )
+      it
+        .execute(
+          dataDeserializer = query.dataDeserializer,
+          dataSerializersModule = query.dataSerializersModule,
+          callerSdkType = query.callerSdkType,
+          fetchPolicy = fetchPolicy,
+        )
+        .toDataSourcePair()
     }
 
   suspend fun <Data, Variables> subscribe(
@@ -45,7 +50,17 @@ internal class QueryManager(private val liveQueries: LiveQueries) {
         dataSerializersModule = query.dataSerializersModule,
         executeQuery = executeQuery,
         callerSdkType = query.callerSdkType,
-        callback = callback,
+        callback = { callback(it.toDataSourcePair()) },
       )
     }
 }
+
+private fun <T> SequencedReference<
+  Result<TaggedReference<com.google.firebase.dataconnect.core.DataSource, T>>
+>
+  .toDataSourcePair(): SequencedReference<Result<DataSourcePair<T>>> = map { result ->
+  result.map { it.toDataSourcePair() }
+}
+
+private fun <T> TaggedReference<com.google.firebase.dataconnect.core.DataSource, T>
+  .toDataSourcePair(): DataSourcePair<T> = DataSourcePair(ref, tag.toDataSourceEnum())
